[WS] Move declarations from AbstractWebsocketConnector to AbstractWebsocketExchange
